+++ 2.java/HRM_VB01/src/com/gmx0807/domain/EmpData.java
@@ -1,70 +1,17 @@
 package com.gmx0807.domain;
 
+import java.util.ArrayList;
+
 //기초 데이터 세팅
 public class EmpData {
-	// emp
-	private int empno;
-	private int mgr;
-	private double sal;
-	private double comm;
-	private int deptno;
-	private String ename;
-	private String job;
-	private String hiredate;
-
-	public EmpData(int empno, String ename, int mgr, String job, double sal, double comm, String hiredate, int deptno) {
-		this.empno = empno;
-		this.mgr = mgr;
-		this.sal = sal;
-		this.comm = comm;
-		this.deptno = deptno;
-		this.ename = ename;
-		this.job = job;
-		this.hiredate = hiredate;
+	public ArrayList<EmpBean> defData() {
+		ArrayList<EmpBean> emp = new ArrayList<EmpBean>();
+		emp.add(new EmpBean(9999, "Smith", 1111, "IT", 3000.0, 300.0, "2015-02-02", 10));
+		emp.add(new EmpBean(1111, "KING", 1111, "President", 5000.0, 800.0, "1996-12-02", 10));
+		emp.add(new EmpBean(2323, "Allen", 9999, "Salesman", 1600.0, 300.0, "2075-03-22", 20));
+		
+		return emp;
 	}
 	
-
-	public void setEmpno(int empno) {
-		this.empno = empno;
-	}
-
-
-	public void setMgr(int mgr) {
-		this.mgr = mgr;
-	}
-
-
-	public void setSal(double sal) {
-		this.sal = sal;
-	}
-
-
-	public void setComm(double comm) {
-		this.comm = comm;
-	}
-
-
-	public void setDeptno(int deptno) {
-		this.deptno = deptno;
-	}
-
-
-	public void setEname(String ename) {
-		this.ename = ename;
-	}
-
-
-	public void setJob(String job) {
-		this.job = job;
-	}
-
-
-	public void setHiredate(String hiredate) {
-		this.hiredate = hiredate;
-	}
-
-
-	public String getData() {
-		return empno + " " + ename + " " + mgr + " " + job + " " + sal + " " + comm + " " + hiredate + " " + deptno; 
-	}
+	
 }
