+++ 2.java/HRM_VB01/src/com/gmx0807/view/SearchEmp.java
@@ -2,7 +2,9 @@ package com.gmx0807.view;
 
 import java.io.BufferedReader;
 import java.io.IOException;
+import java.util.ArrayList;
 
+import com.gmx0807.domain.EmpBean;
 import com.gmx0807.domain.EmpData;
 
 //1.사원정보 조회
@@ -10,9 +12,9 @@ public class SearchEmp {
 	public BufferedReader br;
 	private boolean isStop = false;
 	private String plag = "";
-	private EmpData[] emp;
+	private ArrayList<EmpBean> emp;
 
-	public SearchEmp(BufferedReader br, EmpData[] emp) {
+	public SearchEmp(BufferedReader br, ArrayList<EmpBean> emp) {
 		this.br = br;
 		this.emp = emp;
 	}
@@ -61,6 +63,8 @@ public class SearchEmp {
 			}
 
 		}
+		
+		System.out.println("조회 결과");
 
 		if (plag.equals("a")) {
 			new SearchEmp().getAllEmp(emp);
@@ -72,35 +76,25 @@ public class SearchEmp {
 	}
 
 	// 모든 정보
-	public void getAllEmp(EmpData[] emp) {
-//		System.out.println("사원번호" + "   " + "이름" + "      " + "매니저" + "\t" + "직업" + "\t" + "월급" + "      " + "보너스"
-//				+ "         " + "등록일" + "     " + "근무지번호");
+	public void getAllEmp(ArrayList<EmpBean> emp) {
 
-//		0: empno 1: ename 2: mgr 3: job 4: sal 5: comm 6: hiredate 7:deptno
-		System.out.println("조회 결과");
-
-		for (int i = 0; i < emp.length; i++) {
-			// split data
-			String[] empArr = emp[i].getData().split(" ");
-			// print data
-			for (int j = 0; j < empArr.length; j++) {
-				System.out.print(empArr[j] + "\t");
-			}
-			System.out.println();
+		for (int i = 0; i < emp.size(); i++) {
+			System.out.println(emp.get(i).getData());
 		}
 		System.out.println();
 
 	}
 
 	// 선택된 정보
-	public void getSelectEmp(EmpData[] emp, int sEmp) {
+	public void getSelectEmp(ArrayList<EmpBean> emp, int sEmp) {
 		boolean isfind = false;
-		System.out.println("조회 결과");
+		
+//		0: empno 1: ename 2: mgr 3: job 4: sal 5: comm 6: hiredate 7:deptno
 
-		for (int i = 0; i < emp.length; i++) {
+		for (int i = 0; i < emp.size(); i++) {
 
 			// split data
-			String[] empArr = emp[i].getData().split(" ");
+			String[] empArr = emp.get(i).getData().split(" ");
 
 			// find empno
 			if (sEmp == Integer.parseInt(empArr[0])) {
@@ -117,6 +111,11 @@ public class SearchEmp {
 		}
 		
 		if(!isfind) System.out.println("입력한 사원번호는 없는 번호입니다.");
+		System.out.println();
 		
 	}
 }
+
+
+//System.out.println("사원번호" + "   " + "이름" + "      " + "매니저" + "\t" + "직업" + "\t" + "월급" + "      " + "보너스"
+//+ "         " + "등록일" + "     " + "근무지번호");
